+++ Robust.Shared.CompNetworkGenerator/ComponentNetworkGenerator.cs
@@ -24,14 +24,19 @@ namespace Robust.Shared.CompNetworkGenerator
         private const string GlobalNetEntityName = "global::Robust.Shared.GameObjects.NetEntity";
         private const string GlobalNetEntityNullableName = "global::Robust.Shared.GameObjects.NetEntity?";
 
+        private const string GlobalWeakEntityReferenceName = "global::Robust.Shared.GameObjects.WeakEntityReference";
+        private const string GlobalWeakEntityReferenceNullableName = "global::Robust.Shared.GameObjects.WeakEntityReference?";
+
         private const string GlobalEntityCoordinatesName = "global::Robust.Shared.Map.EntityCoordinates";
         private const string GlobalNullableEntityCoordinatesName = "global::Robust.Shared.Map.EntityCoordinates?";
 
         private const string GlobalEntityUidSetName = "global::System.Collections.Generic.HashSet<global::Robust.Shared.GameObjects.EntityUid>";
         private const string GlobalNetEntityUidSetName = $"global::System.Collections.Generic.HashSet<{GlobalNetEntityName}>";
+        private const string GlobalWeakEntityReferenceSetName = $"global::System.Collections.Generic.HashSet<{GlobalWeakEntityReferenceName}>";
 
         private const string GlobalEntityUidListName = "global::System.Collections.Generic.List<global::Robust.Shared.GameObjects.EntityUid>";
         private const string GlobalNetEntityUidListName = $"global::System.Collections.Generic.List<{GlobalNetEntityName}>";
+        private const string GlobalWeakEntityReferenceListName = $"global::System.Collections.Generic.List<{GlobalWeakEntityReferenceName}>";
 
         private const string GlobalDictionaryName = "global::System.Collections.Generic.Dictionary<TKey, TValue>";
         private const string GlobalHashSetName = "global::System.Collections.Generic.HashSet<T>";
@@ -225,6 +230,28 @@ namespace Robust.Shared.CompNetworkGenerator
                         deltaApply.Add($"fullState.{name} = {name};");
 
                         break;
+                    case GlobalWeakEntityReferenceName:
+                    case GlobalWeakEntityReferenceNullableName:
+                        networkedType = $"NetEntity?";
+
+                        stateFields.Append($@"
+        public {networkedType} {name} = default!;");
+
+                        getField = $"GetNetEntity(Resolve(component.{name}))";
+                        cast = $"(NetEntity?)";
+
+                        handleStateSetters.Append($@"
+            component.{name} = TryGetEntity(state.{name}, out var {name}) ? GetWeakReference({name}.Value) : WeakEntityReference.Invalid;");
+
+                        deltaHandleFields.Append($@"
+                    component.{name} = TryGetEntity({cast} {fieldHandleValue}, out var {name}) ? GetWeakReference({name}.Value) : WeakEntityReference.Invalid;");
+
+                        shallowClone.Append($@"
+                {name} = this.{name},");
+
+                        deltaApply.Add($"fullState.{name} = {name};");
+
+                        break;
                     case GlobalEntityCoordinatesName:
                     case GlobalNullableEntityCoordinatesName:
                         networkedType = $"NetCoordinates{nullableAnnotation}";
@@ -268,6 +295,27 @@ namespace Robust.Shared.CompNetworkGenerator
                         deltaApply.Add($@"fullState.{name} = {name};");
 
                         break;
+                    case GlobalWeakEntityReferenceSetName:
+                        networkedType = GlobalNetEntityUidSetName;
+
+                        stateFields.Append($@"
+                        public {networkedType} {name} = default!;");
+
+                        getField = $"GetNetEntitySet(Resolve(component.{name}))";
+                        cast = $"({GlobalNetEntityUidSetName})";
+
+                        handleStateSetters.Append($@"
+            component.{name} = GetWeakReferenceSet(GetEntitySet(state.{name}));");
+
+                        deltaHandleFields.Append($@"
+                    component.{name} = GetWeakReferenceSet(GetEntitySet({cast} {fieldHandleValue}))");
+
+                        shallowClone.Append($@"
+                {name} = this.{name},");
+
+                        deltaApply.Add($@"fullState.{name} = {name};");
+
+                        break;
                     case GlobalEntityUidListName:
                         networkedType = $"{GlobalNetEntityUidListName}";
 
@@ -282,6 +330,27 @@ namespace Robust.Shared.CompNetworkGenerator
 
                         deltaHandleFields.Append($@"
                     EnsureEntityList<{componentName}>({cast} {fieldHandleValue}, uid, component.{name});");
+
+                        shallowClone.Append($@"
+                {name} = this.{name},");
+
+                        deltaApply.Add($@"fullState.{name} = {name};");
+
+                        break;
+                    case GlobalWeakEntityReferenceListName:
+                        networkedType = $"{GlobalNetEntityUidListName}";
+
+                        stateFields.Append($@"
+                        public {networkedType} {name} = default!;");
+
+                        getField = $"GetNetEntityList(Resolve(component.{name}))";
+                        cast = $"({GlobalNetEntityUidListName})";
+
+                        handleStateSetters.Append($@"
+            component.{name} = GetWeakReferenceList(GetEntityList(state.{name}));");
+
+                        deltaHandleFields.Append($@"
+                    component.{name} = GetWeakReferenceList(GetEntityList({cast} {fieldHandleValue}))");
 
                         shallowClone.Append($@"
                 {name} = this.{name},");